edits3 part1 (doc comments,proxy prep edits, .gitignor edits)

+++ contracts/interfaces/IMLMLevelLogic.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.12;
 
 interface IMLMLevelLogic {
-    function getLevelBySum(uint sum) external view returns (uint);
+    function getLevelBySum(uint256 sum) external view returns (uint256);
 
-    function getPercentByDepth(uint s) external view returns (uint);
+    function getPercentByDepth(uint256 s) external view returns (uint256);
 }
